processing/process_cnpj/requirements.txt
kafka-python==1.4.7
requests==2.22
pydgraph
